s: Replace custom impl TransparentPtrType type
GStringPtr implements many more traits, which could come in handy for users.
It also seems good on principle to lean into GLib instead of forging our own path.

GStringPtr is however not Send.
To deal with this we have to convert to GString before sending across threads,
but this seems like a minor inconvenience and perhaps there is a reason for it.